fix: user addresses endpoints

+++ src/database/migrations/20240131211510-userAddresses.js
@@ -22,11 +22,11 @@ module.exports = {
         allowNull: false,
       },
       lat: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(10, 8),
         allowNull: true,
       },
       lng: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(11, 8),
         allowNull: false,
       },
       userID: {
